+++ plugins/audit/package.json
@@ -13,6 +13,7 @@
     "format": "prettier --single-quote --trailing-comma none --write \"{src,test}/**/*.js\"",
     "lint": "eslint .",
     "prepublish": "npm run build",
+    "coverage": "codecov",
     "commitmsg": "commitlint -e $GIT_PARAMS",
     "release": "standard-version -a -s",
     "type-check": "tsc --noEmit",
@@ -34,27 +35,30 @@
     "request": "2.88.0"
   },
   "devDependencies": {
+    "@types/express": "4.16.1",
+    "@types/request": "2.48.1",
     "@commitlint/cli": "7.2.1",
     "@commitlint/config-conventional": "7.1.2",
     "@commitlint/travis-cli": "7.2.1",
-    "@types/express": "4.16.1",
     "@types/http-errors": "1.6.1",
     "@types/jest": "23.3.13",
+    "@types/memory-fs": "0.3.2",
     "@types/node": "10.12.18",
-    "@types/request": "2.48.1",
     "@typescript-eslint/eslint-plugin": "1.1.0",
     "@verdaccio/babel-preset": "0.0.3",
     "@verdaccio/types": "5.0.0-beta.1",
+    "codecov": "3.0.4",
     "cross-env": "5.2.0",
     "eslint": "5.6.1",
     "eslint-config-google": "0.10.0",
     "eslint-config-prettier": "3.1.0",
     "eslint-plugin-jest": "21.24.1",
     "eslint-plugin-prettier": "3.0.0",
-    "husky": "1.3.1",
+    "husky": "0.14.3",
     "jest": "24.0.0",
-    "lint-staged": "8.1.0",
-    "prettier": "1.15.3",
+    "lint-staged": "7.3.0",
+    "prettier": "1.14.3",
+    "rmdir-sync": "1.0.1",
     "standard-version": "4.4.0",
     "typescript": "3.2.4"
   },
