+++ plugins/audit/test/audit.spec.ts
@@ -2,6 +2,10 @@ import ProxyAudit, { ConfigAudit } from '../src/index';
 
 import { Logger } from '@verdaccio/types';
 
+const config: ConfigAudit = {
+  enabled: true
+};
+
 const logger: Logger = {
   error: e => console.warn(e),
   info: e => console.warn(e),
@@ -14,9 +18,6 @@ const logger: Logger = {
 
 describe('Audit plugin', () => {
   test('should test audit', () => {
-    const config = {
-      enabled: false
-    };
     const audit = new ProxyAudit(config, { logger, config: undefined });
     expect(audit).toBeDefined();
   });
