--- plugins/audit/test/partials/config.ts
@@ -1,7 +1,0 @@
-import { ConfigAudit } from "../../src";
-
-const config: ConfigAudit = {
-  enabled: true
-};
-
-export default config;
